Move sources management out of docker package
